Use autogenerated and correct syscall

+++ src/syscall/syscall_redox_amd64.go
@@ -9,7 +9,7 @@ func setTimespec(sec, nsec int64) Timespec {
 }
 
 func setTimeval(sec, usec int64) Timeval {
-	return Timeval{Sec: sec, Usec: usec}
+	return Timeval{Sec: sec, Usec: int32(usec)}
 }
 
 func (iov *Iovec) SetLen(length int) {
@@ -17,5 +17,5 @@ func (iov *Iovec) SetLen(length int) {
 }
 
 func (cmsg *Cmsghdr) SetLen(length int) {
-	cmsg.Len = uint32(length)
+	cmsg.Len = uint64(length)
 }
